ad-case/technical-solution/backend/requirements.txt
# FastAPI 相关
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pydantic>=2.5.0
pydantic-settings>=2.1.0

# 数据库相关
asyncpg>=0.29.0
psycopg2-binary>=2.9.0  # PostgreSQL 适配器（用于同步操作）

# 向量模型相关（如果需要）
FlagEmbedding>=1.2.0

# 其他工具
python-dotenv>=1.0.0
